Move accumulated y calculations into a separate file
